requirements.txt
mysql-connector-python
fpdf2
jinja2